Add routing and routing test for BlazorCraft
Adds routing to various exercise files and implements corresponding tests to evaluate functionality. New routes and route parameters are created in the exercise files whereas, the test cases have been established to verify if the component has specific routes defined, if the route parameters are defined with the correct name and type, and if the components render correctly.