Add DOM watcher in BlazorCraft web application
Added a new 'domWatcher' function to 'index.html' which sends changes made on the DOM back to Blazor. Updated 'AdvancedCodeSnippet.razor' to include changes related to this DOM watcher addition. It now uses this 'domWatcher' to constantly monitor for changes in the DOM and reflect them on the server side in real-time.