Add exam completion and results sending functionality
A new feature was added to check if all the exams have been successfully completed, this method was included in the TestRunnerService. Moreover, a SendResultsService was implemented to send results to Dropbox. A corresponding button was added in the MainLayout, which gets activated once all exams are completed. In addition, the Dropbox.Api was added to the project dependencies.