Update component references in instruction texts
The code changes update the way components are referenced in instruction texts across multiple pages. Previously, C# naming was used which could be confusing for the user. Now, the 'ComponentFileWithPath' structure is being used that provides a clearer and more understandable component reference. This way, users can more easily locate and modify the correct component as instructed in the exercises.