Update CSS and enable dark theme in application
The commit updates the CSS including changes to the 'MudThemeProvider' and addition of 'prism_dark.css'. The stick panel header's background color in 'app.css' is changed to use variable '--mud-palette-background'. In addition, the Prism CSS has been switched to use a dark theme 'prism_dark.css'. 'TestRunner.razor' has been modified to return the default color value instead of dark when the value is null. Finally, the application is set to dark mode by default.