Update README with additional prerequisites and long path support instructions
A section has been added to the README which includes additional prerequisites needed to run the project. This now includes Windows 10 version 1607 (Anniversary Update) or later, and the need for administrative privileges on the machine. Also, guidance on enabling long path support in Windows has been introduced for working with deeply nested files and directories.